Use one file each for wrapper type definitions and tests.
Matches the convention of this project.
